Added login verify option
localhost:3000/auth/verify will get a request from useEffect() hook from header.jsx and returns a response to FE that says whether User is logged in or not. So using conditional rendering, FE will show whether it needs to show Login or Logout in the header navigation bar.
For proper functionality Redux will be added later.